Cast parent_query_id to int64_t before appending
Just for consistency with the other cast... I don't know why that's being done in the first place...

+++ src/export/stats_exporter.cc
@@ -149,7 +149,7 @@ void ExportEventStats(const std::vector<PschEvent>& events, StatsExporter* expor
     col_username->Append(std::string(ev.username, ev.username_len));
     col_pid->Append(ev.pid);
     col_query_id->Append(static_cast<int64_t>(ev.queryid));
-    col_parent_query_id->Append(ev.parent_query_id);
+    col_parent_query_id->Append(static_cast<int64_t>(ev.parent_query_id));
     col_cmd_type->Append(CmdTypeToString(ev.cmd_type));
     col_rows->Append(ev.rows);
